+++ src/test/java/jscover/report/MainTest.java
@@ -478,10 +478,9 @@ public class MainTest {
         verify(ioUtils).copy("mergedJSON", mergedJson);
 
         //Verify src copying
-        File src1 = new File(dir1, jscover.Main.reportSrcSubDir);
         File src2 = new File(dir2, jscover.Main.reportSrcSubDir);
         File srcDest = new File(destDir, jscover.Main.reportSrcSubDir);
-        verify(ioUtils).copyDir(src1, srcDest);
+        verify(ioUtils).copyDir(dir1, destDir);
         verify(ioUtils).copyDir(src2, srcDest);
     }
 }