+++ History.md
@@ -1,11 +1,12 @@
 0.2.0-RC1 / 2012-??-??
 ==================
-  * Alter JSONDataMerger.mergeJSONCoverageStrings to accept and array of coverage strings to merge
+  * Add report merging command line interface
   * Improve unloaded JavaScript parsing error logging issue https://github.com/tntim96/JSCover/issues/27
-  * Internal (NB These change are not backwards compatible):
+  * Internal:
         Load original JS source from file-system instead of JSON for report viewing
         Remove source from JSON
         Alter _$jscoverage object and JSON to elegantly accommodate branch (and other) properties
+        Alter JSONDataMerger.mergeJSONCoverageStrings to accept and array of coverage strings to merge
 
 0.1.1 / 2012-12-02
 ==================